Make `format` smart about multiline messages (#1)
Before, there were three cases, depending on the format used:

  * `endgroup`:                  no argument
  * `command`, `group`:          one argument
  * `debug`, `error`, `warning`: one or more arguments

Each argument became one line, prepended with a formatting command.

One problem with this was that one might do e.g.

    log(format("error")("A line\nAnother line"));

expecting

    ##[error]A line
    ##[error]Another line

but actually getting

    ##[error]A line
    Another line

in which case only the first line would be colored red in the log.

Now, there are two cases:

  * `endgroup`:        no argument
  * All other formats: zero or more arguments

Each of the arguments is now split at newlines; then each line is
formatted individually. It is thus on the user to ensure that they don't
pass a multiline argument if they don't want multiple formatted lines.